+++ synthetic_data_benchmark/evaluator/summary.py
@@ -20,11 +20,19 @@ def method_name_order(name):
         return 0
     if 'clbn' in name.lower():
         return 1
-    if 'vae' in name.lower():
+    if 'privbn' in name.lower():
         return 2
-    if 'gan' in name.lower():
+    if 'medgan' in name.lower():
         return 3
-    return 4
+    if 'veegan' in name.lower():
+        return 4
+    if 'tablegan' in name.lower():
+        return 5
+    if 'tvae' in name.lower():
+        return 8
+    if 'tgan' in name.lower():
+        return 9
+    return 6
 
 def coverage(datasets, results):
     ticks = []
